Less harsh attribute lookup
The uPy convention is that loading an attribute shouldn't raise an AttributeError
for non-existing attributes but instead return MP_OBJ_NULL to indicate the
attribute was not found. See mp_builtin_hasattr for instance where this is
mentioned explicitly.
Fix ClassWrapper::load_attr() accordingly so calls like hasattr work properly
instaid of raising an AttributeError, and add tests covering this behaviour.