Improve type check used when converting uPy objects to native types
The trick with the cookie only works if actual objects are passed in: when
passing for instance an integer to AsNativeObjChecked it would just segfault
due to trying to access memory at low locations.
Moreover the error raised doesn't contain a lot of useful information.
But even if the cookie matches it doesn't mean the type matches as that
isn't checked at all so that's another receipe for segfaults.

Replace this with an actual proper type check and raise a TypeError with
source and target type names to ease debugging.