Allow relaxing the type check introduced in [e8081fa]
Applications which register C++ functions like 'void foo(Base* p)' and
want to pass instances of derived classes to that don't work with the type
check as it requires an exact match and subclasses aren't considered.
Introduce a macro which can turn off the full type check, though be aware
that this will cause havoc when passing unrelated types. (to fix that
we'd probably have to keep a map of derived types or so)