Produce a proper error message when trying to convert a const reference
We don't convert from const reference arguments to ClassWrapper instances
since we either cannot guarantee the object really won't be modified
(since python has no notion of const) or else would make a copy (which
is sort of unexpected and confusing).
However when writing code which tries this anyway this would just lead
to sometimes obscure error messages like
"cannot convert from const A to A&"
so introduce a static_assert with a clear message to deal with this.