Support constructor functions returning shared_ptr
Allow custom functions for constructing an instance (i.e. used instead
of the class constructor) to return a shared_ptr instead of requiring
a bare pointer to the instance.
This is much cleaner because it doesn't force the user to manually write
new() but enables using make_shared() instead with all benefits that has,
and allows that in the core code as well.