Add conversion of None to nullptr
A native argument which is a T* or an std::shared_ptr< T > will now
be passed as nullptr when the Python code passes None. Only when taking
a T& a TypeError will be raised. This is more consistent in general
since after all pointer can be null in C++. It does however mean any
code which takes a pointer argument and blindly assumes it's not null
(which it obviously shouldn't ever do anyway) will break.