tests: Rename C user module file
Building the C user module for the tests for e.g. the ESP32 port
requires some build customisations to the ESP32 Makefile to make
it compile C++ source files. To make this easier rename the .c file
so it has a different name than the .cpp file otherwise the
resulting object file (module.o) is the same for both.