Rewrite configuration options
Improve how the configuration of features via the preprocessor happens
by using one single header which is clearer and easier to maintain, and
by not using plain #ifdef but #if for consistency.
This last change remains backwards compatible because the 2 affected
options also are the ones which use a double negation '#ifndef NO_XXX'
so a 'positive' version is introduced next to it.