classwrapper: Decouple construction from type registration
To allow more use cases in the future don't restrict to initializing
and storing type and locals in a dict, but split the constructor into
one which initializes and delegate constructors which store the needed
parts in a dict. Also enforce correct usage (when called multiple times
flags must be the same) while allowing the same type with different names.